requirements.txt
pandas
numpy
openai-whisper
schedule
spacy
# Para NLP em português
https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.7.0/pt_core_news_sm-3.7.0-py3-none-any.whl